recommendation-service/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.31
pydantic==2.8.2
aiokafka==0.10.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
jsonschema
# psycopg2==2.9.10
